+++ test/conftest.py
@@ -1,0 +1,18 @@
+import os
+import tempfile
+import pytest
+
+REMOVE_DB = False
+
+
+@pytest.fixture
+def db_path():
+    db_file_location = os.path.join(tempfile.gettempdir(), "abc_unittest.db")
+    db = "sqlite:///" + db_file_location
+    yield db
+    if REMOVE_DB:
+        try:
+            if REMOVE_DB:
+                os.remove(db_file_location)
+        except FileNotFoundError:
+            pass